refactor plantHardinessZone
removed userSelectRegion redux, update save point to redux in selectedToEditSite useEffect
