+++ src/pages/CropSidebar/PlantHardinessZone/PlantHardinessZone.js
@@ -4,20 +4,23 @@ import {
 } from '@mui/material';
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { userSelectRegion } from '../../../reduxStore/userSlice';
-import statesLatLongDict from '../../../shared/stateslatlongdict';
-// import { useSelector } from 'react-redux';
+import { updateRegion } from '../../../reduxStore/mapSlice';
 
-const PlantHardinessZone = ({
-  regionShorthand,
-  setRegionShorthand,
-  regionsRedux,
-  councilLabelRedux,
-  regionToggleRedux = true,
-}) => {
+const PlantHardinessZone = () => {
   const dispatchRedux = useDispatch();
-  const markersRedux = useSelector((stateRedux) => stateRedux.addressData.markers);
-  const stateLabelRedux = useSelector((stateRedux) => stateRedux.mapData.stateLabel);
+  const regionsRedux = useSelector((stateRedux) => stateRedux.mapData.regions);
+  const councilLabelRedux = useSelector((stateRedux) => stateRedux.mapData.councilLabel);
+  const regionToggleRedux = useSelector((stateRedux) => stateRedux.sharedData.regionToggle);
+  const regionShorthandRedux = useSelector((stateRedux) => stateRedux.mapData.regionShorthand);
+
+  const updateRegionRedux = (e) => {
+    const selectedRegion = regionsRedux.filter((region) => region.shorthand === e.target.value)[0];
+    localStorage.setItem('regionId', selectedRegion.id);
+    dispatchRedux(updateRegion({
+      regionId: selectedRegion.id ?? '',
+      regionShorthand: selectedRegion.shorthand ?? '',
+    }));
+  };
 
   const menuProps = {
     PaperProps: {
@@ -67,20 +70,9 @@ const PlantHardinessZone = ({
         },
       }}
       MenuProps={menuProps}
-      onChange={(e) => {
-        setRegionShorthand(e.target.value);
-        if (markersRedux
-          && (markersRedux[0][1] === statesLatLongDict[stateLabelRedux][1]
-         && markersRedux[0][0] === statesLatLongDict[stateLabelRedux][0])) {
-          // set redux to true (means user select another region while didn't change address)
-          dispatchRedux(userSelectRegion(true));
-        } else {
-          // if user changed marker, not set redux
-          dispatchRedux(userSelectRegion(false));
-        }
-      }}
-      value={regionShorthand || ''}
-      error={!regionShorthand}
+      onChange={updateRegionRedux}
+      value={regionShorthandRedux || ''}
+      error={!regionShorthandRedux}
     >
 
       {regionsRedux?.length > 0 && regionsRedux.map((region, i) => (
@@ -109,7 +101,7 @@ const PlantHardinessZone = ({
               {councilLabelRedux === 'Midwest Cover Crops Council' ? 'COUNTY' : 'ZONE'}
             </InputLabel>
             {plantHardinessZone()}
-            {!regionShorthand
+            {!regionShorthandRedux
             && (
             <Typography variant="body2" align="center" color="error" gutterBottom>
               {`Please Select a ${councilLabelRedux === 'Midwest Cover Crops Council' ? 'County' : 'Zone'}`}
